+++ SeriesReduction.py
@@ -118,7 +118,7 @@ class MatrixSolution:
             obsThree = self.balanceReadings[i][2]
             obsFour = self.balanceReadings[i][3]
 
-            swDensityAdjusted = self.swDensity / (1 + self.swCCE * ((np.longdouble(self.environmentals[i][0]) - np.longdouble(self.envCorrections[0])) - self.referenceTemperature))
+            swDensityAdjusted = self.swDensity / (1 + self.swCCE * ((np.float64(self.environmentals[i][0]) - np.float64(self.envCorrections[0])) - self.referenceTemperature))
 
             airDensity = calculateAirDensity(\
                 self.environmentals[i][0], self.envCorrections[0], self.environmentals[i][1], self.envCorrections[1], self.environmentals[i][2], self.envCorrections[2])
@@ -176,7 +176,7 @@ class MatrixSolution:
             #Adjust densities for lab temperature for each observation:
             adjustedDensities = []
             for j in range(self.positions):
-                adjustedDensities.append(self.weightDensities[j] / (1 + self.weightCCEs[j] * ((np.longdouble(self.environmentals[i][0]) - np.longdouble(self.envCorrections[0])) - self.referenceTemperature)))
+                adjustedDensities.append(self.weightDensities[j] / (1 + self.weightCCEs[j] * ((np.float64(self.environmentals[i][0]) - np.float64(self.envCorrections[0])) - self.referenceTemperature)))
             
             #Estimate Mass1Sum, Mass2Sum and effective densities for ABC using estimateMasses:
             designLine = self.designMatrix[i:i+1] #Get sigle line of design matrix as an array
